pyproject.toml
[tool.poetry]
name = "reforma-tributaria-api"
version = "0.1.0"
description = "API FastAPI com autenticação e autorização (administrator/client) em arquitetura limpa."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"

# --- Core Web Framework ---
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }

# --- ORM e Banco de Dados ---
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
mysql-connector-python = "^9.1.0"  # troque para psycopg2-binary se usar PostgreSQL

# --- Segurança e Autenticação ---
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
bcrypt = "4.0.1"                 # <<< FIX contra o erro "__about__" nas versões novas
python-multipart = "^0.0.9"        # Necessário para /auth/login com form-data

# --- Validação e Configuração ---
pydantic = "^2.9.0"
pydantic-settings = "^2.3.3"
python-dotenv = "^1.0.1"
email-validator = "^2.2.0"

# --- Utilitários e Logs ---
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
ruff = "^0.6.4"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]  # já coberto pelo Black

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


